fix: ExceptionHandler 우선순위 차이에 의한 버그 핸들러 통합으로 해결
