--- src/main/java/gift/config/AuthenticationExceptionHandler.java
@@ -1,34 +1,0 @@
-package gift.config;
-
-import gift.exception.EmailAlreadyExistsException;
-import gift.exception.InvalidLoginException;
-import gift.exception.InvalidRegisterParameterException;
-import gift.exception.UnauthenticatedException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class AuthenticationExceptionHandler {
-
-    @ExceptionHandler(EmailAlreadyExistsException.class)
-    public ResponseEntity<String> handlerException(EmailAlreadyExistsException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.CONFLICT);
-    }
-
-    @ExceptionHandler(InvalidLoginException.class)
-    public ResponseEntity<String> handlerException(InvalidLoginException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.FORBIDDEN);
-    }
-
-    @ExceptionHandler(InvalidRegisterParameterException.class)
-    public ResponseEntity<String> handlerException(InvalidRegisterParameterException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(UnauthenticatedException.class)
-    public ResponseEntity<String> handlerException(UnauthenticatedException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.UNAUTHORIZED);
-    }
-}
